requirements.txt
alembic==1.13.1
blinker==1.8.1
click==8.1.7
dnspython==2.6.1
email_validator==2.1.1
Faker==25.0.1
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WhooshAlchemy==0.56
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
lorem==0.1.1
Mako==1.3.3
MarkupSafe==2.1.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
SQLAlchemy==2.0.30
typing_extensions==4.11.0
Werkzeug==3.0.3
Whoosh==2.7.4
WTForms==3.1.2
